requirements.txt
streamlit==1.33.0
PyMuPDF==1.23.7
pytesseract==0.3.10
Pillow==10.3.0
google-generativeai==0.4.1
